// app/frontend/src/app/models/zahtev.ts
export class Zahtev{
  kor_ime: string = '';
  lozinka: string = '';
  ime: string = '';
  prezime: string = '';
  mejl: string = '';
  tip: string = '';
  pol: string = '';
  adresa: string = '';
  kontakt: string = '';
  brKartice: string = '';
  profilna: string = '';
  status: string = '';
  idZ: number= 0;
}